projekt_quiz/backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
nltk>=3.8.1 